Select Alliance Portfolios of Germany and Russia, 1914

A simple example of alliance portfolios of Germany and Russia, by way of Signorino and Ritter's (1999) example.

Usage

gmyrus14

Format

A data frame with 20 observations on the following 4 variables.

state

a three-character code indicating a state

syscap

the capabilities of the state, as a potential weight

gmy

the alliance commitment for Germany with the state identified in the state column

rus

the alliance commitment for Russia with the state identified in the state column

Details

The data come by way of Signorino and Ritter (1999, Table 6).

References

Signorino, Curtis S. and Jeffrey M. Ritter. "Tau-b or Not Tau-B: Measuring the Similarity of Foreign Policy Positions." International Studies Quarterly 43(1): 115–44.